% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUtils.R
\name{depend}
\alias{depend}
\title{Depend on a Javascript library available at a URL}
\usage{
depend(url, name = NULL, version = NULL)
}
\arguments{
\item{url}{URL of the library.}

\item{name}{Name of library. \code{htmltools} will include the library exactly once in the document, using the latest mentioned version.}

\item{version}{Version of the library, should be interpretable by \code{numeric_version()}.}
}
\value{
An empty \code{htmltools::tagList} that depends on the specified URL. Creates the dependency when it is printed in an rmarkdown document.
}
\description{
Include a Javascript dependency, for example in an rmarkdown document.
}
\details{
Downloads the script at the given URL if it is not aleady cached. Returns an empty HTML object that depends on the library. \code{name} and \code{version} are used by \code{htmltools} to decide which version of a library to include. If several versions are depended on, the latest version is chosen. Of course there is the possibility of subtle bugs because of this! If no name is given, the hash of the url is used.

If you're using the same package as some widget, make sure you use the same name (\code{htmltools::findDependencies(someWidget)}), or just rely on the widget for the dependency.

The file cache directory is chosen using \code{rappdirs::user_cache_dir("jsandr")}.
}
\examples{
depend("https://unpkg.com/svd-js@1.1.1/build-umd/svd-js.min.js")

}
