% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUtils.R
\name{dependD3}
\alias{dependD3}
\title{Depend on D3}
\usage{
dependD3(...)
}
\arguments{
\item{...}{Parameters for \code{r2d3::html_dependencies_d3}.}
}
\description{
Use a version of D3 included in r2d3.
}
\examples{
dependD3()

}
