% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUtils.R
\name{let}
\alias{let}
\alias{let1}
\title{Send data to Javascript}
\usage{
let(...)

let1(...)
}
\arguments{
\item{...}{Data to send to Javascript. Argument names become variable names.}
}
\description{
Make data available in Javascript. Data is encoded using jsonlite.
}
\details{
The let1 variant is for setting single numbers, which are otherwise encoded as arrays of length 1.
}
\examples{
let(x=c(1,2,3,4), y=c(5,6,7,8))

let1(n=4)

}
