% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talkToPlotly.R
\name{talkToPlotly}
\alias{talkToPlotly}
\title{Improved crosstalk behaviour with plotly widget}
\usage{
talkToPlotly(
  plotlyObject,
  dragmode = "select",
  webGL = TRUE,
  unselectedColor = "#bbbbbb",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Try to provide standard behaviour for selections and filters.
}
